+++ src/main/java/dev/willyelton/crystal_tools/network/ToolAttributePacket.java
@@ -5,6 +5,7 @@ import dev.willyelton.crystal_tools.utils.NBTUtils;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraftforge.network.NetworkEvent;
 
@@ -41,20 +42,30 @@ public class ToolAttributePacket {
             ItemStack heldTool = ItemStackUtils.getHeldLevelableTool(playerEntity);
 
             if (!heldTool.isEmpty()) {
-                // need to do something about this maybe
-                if (msg.key.equals("silk_touch_bonus")) {
-                    heldTool.enchant(Enchantments.SILK_TOUCH, 1);
-                } else if (msg.key.equals("fortune_bonus")) {
-                    NBTUtils.removeEnchantments(heldTool);
-                    heldTool.enchant(Enchantments.BLOCK_FORTUNE, (int) msg.value);
+                Enchantment enchantment = enchantmentFromString(msg.key);
+
+                if (enchantment != null) {
+                    heldTool.enchant(enchantment, (int) msg.value);
                 } else if (msg.key.equals("auto_repair")) {
                     NBTUtils.setValue(heldTool, msg.key, true);
                     NBTUtils.addValueToTag(heldTool, "auto_repair_amount", 1);
-                } else if(msg.key.equals("looting_bonus")) {
-                    heldTool.enchant(Enchantments.MOB_LOOTING, (int) msg.value);
                 } else {
                     NBTUtils.addValueToTag(heldTool, msg.key, msg.value);
                 }
+
+//                if (msg.key.equals("silk_touch_bonus")) {
+//                    heldTool.enchant(Enchantments.SILK_TOUCH, 1);
+//                } else if (msg.key.equals("fortune_bonus")) {
+//                    NBTUtils.removeEnchantments(heldTool);
+//                    heldTool.enchant(Enchantments.BLOCK_FORTUNE, (int) msg.value);
+//                } else if (msg.key.equals("auto_repair")) {
+//                    NBTUtils.setValue(heldTool, msg.key, true);
+//                    NBTUtils.addValueToTag(heldTool, "auto_repair_amount", 1);
+//                } else if(msg.key.equals("looting_bonus")) {
+//                    heldTool.enchant(Enchantments.MOB_LOOTING, (int) msg.value);
+//                } else {
+//                    NBTUtils.addValueToTag(heldTool, msg.key, msg.value);
+//                }
 
                 // update the skill points array
                 if (msg.id != -1) {
@@ -62,5 +73,21 @@ public class ToolAttributePacket {
                 }
             }
         }
+    }
+
+    private static Enchantment enchantmentFromString(String string) {
+        return switch (string) {
+            case "silk_touch_bonus" -> Enchantments.SILK_TOUCH;
+            case "fortune_bonus" -> Enchantments.BLOCK_FORTUNE;
+            case "looting_bonus" -> Enchantments.MOB_LOOTING;
+            case "protection_bonus" -> Enchantments.ALL_DAMAGE_PROTECTION;
+            case "fire_protection_bonus" -> Enchantments.FIRE_PROTECTION;
+            case "blast_protection_bonus" -> Enchantments.BLAST_PROTECTION;
+            case "projectile_protection_bonus" -> Enchantments.PROJECTILE_PROTECTION;
+            case "feather_falling_bonus" -> Enchantments.FALL_PROTECTION;
+            case "aqua_affinity_bonus" -> Enchantments.AQUA_AFFINITY;
+            case "respiration_bonus" -> Enchantments.RESPIRATION;
+            default -> null;
+        };
     }
 }