+++ changelog.md
@@ -1,3 +1,7 @@
+----------------------- 0.4.1 Changes -----------------------
+### Changes
+- Changed the crystal bow texture to a custom one
+
 ----------------------- 0.4.0 Changes -----------------------
 ### Additions
 - Added the crystal Bow